Atualize códigos e correção de algumas falhas

+++ Projetos/first-project/src/main/java/com/example_first_project/repository/PrioridadeRepository.java
@@ -3,5 +3,5 @@ package com.example_first_project.repository;
 import com.example_first_project.entity.PrioridadeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PrioridadeRepository extends JpaRepository<PrioridadeEntity, Integer> {
+public interface PrioridadeRepository extends JpaRepository<PrioridadeEntity, Long> {
 }
