+++ Projetos/first-project/src/main/java/com/example_first_project/repository/StatusRepository.java
@@ -3,5 +3,5 @@ package com.example_first_project.repository;
 import com.example_first_project.entity.StatusEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StatusRepository extends JpaRepository<StatusEntity, Integer> {
+public interface StatusRepository extends JpaRepository<StatusEntity, Long> {
 }
